+++ R/ggscatter.R
@@ -57,6 +57,8 @@ NULL
 #'   text labels or not.
 #' @param label.rectangle logical value. If TRUE, add rectangle underneath the
 #'   text, making it easier to read.
+#' @param parse If \code{TRUE}, the labels will be parsed into expressions and
+#'   displayed as described in \code{?plotmath}.
 #' @param cor.coef logical value. If TRUE, correlation coefficient with the
 #'   p-value will be added to the plot.
 #' @param cor.coeff.args a list of arguments to pass to the function
@@ -153,6 +155,7 @@ ggscatter <- function(data, x, y, combine = FALSE, merge = FALSE,
                       star.plot = FALSE, star.plot.lty = 1, star.plot.lwd = NULL,
                       label = NULL,  font.label = c(12, "plain"), font.family = "",
                       label.select = NULL, repel = FALSE, label.rectangle = FALSE,
+                      parse = FALSE,
                       cor.coef = FALSE, cor.coeff.args = list(), cor.method = "pearson", cor.coef.coord = c(NULL, NULL), cor.coef.size = 4,
                       ggp = NULL, show.legend.text = NA,
                       ggtheme = theme_pubr(),
@@ -176,6 +179,7 @@ ggscatter <- function(data, x, y, combine = FALSE, merge = FALSE,
     star.plot = star.plot, star.plot.lty = star.plot.lty, star.plot.lwd = star.plot.lwd,
     label = label, font.label = font.label, font.family = font.family,
     label.select = label.select, repel = repel, label.rectangle = label.rectangle,
+    parse = parse,
     cor.coef = cor.coef, cor.coeff.args = cor.coeff.args, cor.method = cor.method,
     cor.coef.coord = cor.coef.coord, cor.coef.size = cor.coef.size,
     ggp = ggp, show.legend.text = show.legend.text, ggtheme = ggtheme, ...)
@@ -221,6 +225,7 @@ ggscatter_core <- function(data, x, y,
                       star.plot = FALSE, star.plot.lty = 1, star.plot.lwd = NULL,
                       label = NULL,  font.label = c(12, "plain"), font.family = "",
                       label.select = NULL, repel = FALSE, label.rectangle = FALSE,
+                      parse = FALSE,
                       cor.coef = FALSE, cor.coeff.args = list(), cor.method = "pearson", cor.coef.coord = c(NULL, NULL), cor.coef.size = 4,
                       ggp = NULL, show.legend.text = NA,
                       ggtheme = theme_classic(),
@@ -345,6 +350,7 @@ ggscatter_core <- function(data, x, y,
       if(label.rectangle) ggfunc <- ggrepel::geom_label_repel
         p <- p + .geom_exec(ggfunc, data = lab_data, x = x, y = y,
                           label = label, fontface = font.label$face,
+                          parse = parse,
                           size = font.label$size/3, color = font.label$color,
                           alpha = alpha, family = font.family,
                           box.padding = unit(0.35, "lines"),
@@ -357,9 +363,10 @@ ggscatter_core <- function(data, x, y,
       if(label.rectangle) {
         ggfunc <- geom_label
         vjust <- -0.4
-        }
+      }
       p <- p + .geom_exec(ggfunc, data = lab_data, x = x, y = y, color = color,
                           label = label, fontface = font.label$face, family = font.family,
+                          parse = parse,
                           size = font.label$size/3, color = font.label$color,
                           vjust = vjust, alpha = alpha, show.legend = show.legend.text)
     }