+++ R/ggtext.R
@@ -11,10 +11,12 @@ NULL
 #' @param size text font size.
 #' @param face text font style. Allowed values are one of c("plain", "bold",
 #'   "italic", "bold.italic").
+#' @param parse If \code{TRUE}, the labels will be parsed into expressions and
+#'   displayed as described in \code{?plotmath}.
 #' @param family character vector specifying font family.
-#' @param show.legend logical. Should text be included in the legends? NA,
-#'   the default, includes if any aesthetics are mapped. FALSE never includes,
-#'   and TRUE always includes.
+#' @param show.legend logical. Should text be included in the legends? NA, the
+#'   default, includes if any aesthetics are mapped. FALSE never includes, and
+#'   TRUE always includes.
 #' @param label.select can be of two formats: \itemize{ \item a character vector
 #'   specifying some labels to show. \item a list containing one or the
 #'   combination of the following components: \itemize{ \item \code{top.up} and
@@ -65,6 +67,7 @@ ggtext <- function(data, x = NULL, y = NULL, label = NULL,
                   color = "black",  palette = NULL,
                   size = 11,  face = "plain", family = "", show.legend = NA,
                   label.select = NULL, repel = FALSE, label.rectangle = FALSE,
+                  parse = FALSE,
                   grouping.vars = NULL,
                   position = "identity",
                   ggp = NULL, ggtheme = theme_pubr(),
@@ -133,7 +136,7 @@ ggtext <- function(data, x = NULL, y = NULL, label = NULL,
                         label = label, fontface = face,
                         family = family, show.legend = show.legend,
                         size = size/3, color = color,
-                        alpha = alpha,
+                        alpha = alpha, parse = parse,
                         box.padding = unit(0.35, "lines"),
                         point.padding = unit(0.3, "lines"),
                         force = 1, segment.size = 0.2, seed = 123)
@@ -150,7 +153,7 @@ ggtext <- function(data, x = NULL, y = NULL, label = NULL,
     if(!is.null(.dots$hjust)) hjust <- .dots$hjust
     p <- p + geom_exec(ggfunc, data = lab_data, x = x, y = y, color = color,
                         label = label, fontface = face, family = family, show.legend = show.legend,
-                        size = size/3, color = color,
+                        size = size/3, color = color, parse = parse,
                         vjust = vjust, hjust = hjust, alpha = alpha, position = position)
 
   }