+++ R/utilities.R
@@ -840,6 +840,7 @@ p
                      add = "none", add.params = list(),
                      label = NULL, font.label = list(size = 11, color = "black"),
                      label.select = NULL, repel = FALSE, label.rectangle = FALSE,
+                     parse = FALSE,
                      ggtheme = theme_pubr(),
                      fun_name = "", group = 1, # used only by ggline
                      show.legend.text = NA,
@@ -975,7 +976,9 @@ p
     label.opts <- font.label %>%
       .add_item(data = data, x = opts$x, y = opts$y,
                 label = label, label.select = label.select,
-                repel = repel, label.rectangle = label.rectangle, ggtheme = NULL,
+                repel = repel, label.rectangle = label.rectangle,
+                parse = parse,
+                ggtheme = NULL,
                 grouping.vars = grouping.vars, facet.by = facet.by, position = geom.text.position,
                 show.legend = show.legend.text)
     p <- purrr::map(p,